+++ composer.json
@@ -38,7 +38,7 @@
         "php": ">=8.2",
         "composer-runtime-api": "^2.0",
         "dragonmantank/cron-expression": "^3.4.0",
-        "laravel/serializable-closure": ">=2.0 <2.0.9",
+        "laravel/serializable-closure": "^2.0",
         "psr/log": "^2.0 || ^3.0",
         "symfony/config": "^6.4.25 || ^7.4.0 || ^8.0.0",
         "symfony/console": "^6.4.25 || ^7.4.0 || ^8.0.0",
@@ -61,6 +61,9 @@
         "phpunit/phpunit": "10.5.60",
         "symfony/error-handler": "^6.4.25 || ^7.4.0 || ^8.0.0",
         "symfony/phpunit-bridge": "^6.4.25 || ^7.4.0 || ^8.0.0"
+    },
+    "conflict": {
+        "laravel/serializable-closure": ">=2.0.9"
     },
     "minimum-stability": "beta",
     "prefer-stable": true,
