+++ tests/Unit/Service/LaravelClosureSerializerTest.php
@@ -5,6 +5,8 @@ declare(strict_types=1);
 namespace Crunz\Tests\Unit\Service;
 
 use Crunz\Infrastructure\Laravel\LaravelClosureSerializer;
+use Crunz\Tests\TestCase\SerializableTaskRunnerStub;
+use Crunz\Tests\TestCase\TaskRunnerStub;
 use Crunz\Tests\TestCase\UnitTestCase;
 
 final class LaravelClosureSerializerTest extends UnitTestCase
@@ -62,7 +64,7 @@ final class LaravelClosureSerializerTest extends UnitTestCase
      */
     public function test_serialize_closure_bound_to_object_with_serialize_and_closure_properties(): void
     {
-        $runner = new TaskRunnerWithSerializeStub();
+        $runner = new SerializableTaskRunnerStub();
         $closure = $runner->createTask();
 
         $serialized = $this->serializer->serialize($closure);
@@ -78,59 +80,5 @@ final class LaravelClosureSerializerTest extends UnitTestCase
         $code = $this->serializer->closureCode($testClosure);
 
         self::assertSame('static fn (): \stdClass => new \stdClass()', $code);
-    }
-}
-
-/**
- * @internal
- */
-class TaskRunnerStub
-{
-    public string $taskName = 'daily-report';
-    /** @var \Closure[] */
-    public array $filters = [];
-
-    public function createTask(): \Closure
-    {
-        $this->filters[] = static function (): bool { return true; };
-
-        return function (): string {
-            return "running {$this->taskName}";
-        };
-    }
-}
-
-/**
- * @internal
- */
-class TaskRunnerWithSerializeStub
-{
-    public string $taskName = 'daily-report';
-    /** @var \Closure[] */
-    public array $filters = [];
-
-    /** @return array{taskName: string, filters: array<\Closure>} */
-    public function __serialize(): array
-    {
-        return [
-            'taskName' => $this->taskName,
-            'filters' => $this->filters,
-        ];
-    }
-
-    /** @param array{taskName: string, filters: array<\Closure>} $data */
-    public function __unserialize(array $data): void
-    {
-        $this->taskName = $data['taskName'];
-        $this->filters = $data['filters'];
-    }
-
-    public function createTask(): \Closure
-    {
-        $this->filters[] = static function (): bool { return true; };
-
-        return function (): string {
-            return "running {$this->taskName}";
-        };
     }
 }
